+++ src/renderer/component/fileDownloadLink/index.js
@@ -9,6 +9,7 @@ import { makeSelectCostInfoForUri } from 'redux/selectors/cost_info';
 import { doFetchAvailability } from 'redux/actions/availability';
 import { doOpenFileInShell } from 'redux/actions/file_info';
 import { doPurchaseUri, doStartDownload } from 'redux/actions/content';
+import { setVideoPause } from 'redux/actions/video';
 import FileDownloadLink from './view';
 
 const select = (state, props) => ({
@@ -24,6 +25,7 @@ const perform = dispatch => ({
   openInShell: path => dispatch(doOpenFileInShell(path)),
   purchaseUri: uri => dispatch(doPurchaseUri(uri)),
   restartDownload: (uri, outpoint) => dispatch(doStartDownload(uri, outpoint)),
+  setVideoPause: val => dispatch(setVideoPause(val)),
 });
 
 export default connect(select, perform)(FileDownloadLink);