+++ src/renderer/component/video/index.js
@@ -4,6 +4,7 @@ import { doChangeVolume } from 'redux/actions/app';
 import { selectVolume } from 'redux/selectors/app';
 import { doPlayUri, doSetPlayingUri } from 'redux/actions/content';
 import { makeSelectMetadataForUri, makeSelectContentTypeForUri } from 'redux/selectors/claims';
+import { setVideoPause } from 'redux/actions/video';
 import {
   makeSelectFileInfoForUri,
   makeSelectLoadingForUri,
@@ -11,6 +12,7 @@ import {
 } from 'redux/selectors/file_info';
 import { makeSelectCostInfoForUri } from 'redux/selectors/cost_info';
 import { selectShowNsfw } from 'redux/selectors/settings';
+import { selectVideoPause } from 'redux/selectors/video';
 import Video from './view';
 import { selectPlayingUri } from 'redux/selectors/content';
 
@@ -24,12 +26,14 @@ const select = (state, props) => ({
   playingUri: selectPlayingUri(state),
   contentType: makeSelectContentTypeForUri(props.uri)(state),
   volume: selectVolume(state),
+  videoPause: selectVideoPause(state),
 });
 
 const perform = dispatch => ({
   play: uri => dispatch(doPlayUri(uri)),
   cancelPlay: () => dispatch(doSetPlayingUri(null)),
   changeVolume: volume => dispatch(doChangeVolume(volume)),
+  setVideoPause: val => dispatch(setVideoPause(val)),
 });
 
 export default connect(select, perform)(Video);