+++ src/renderer/redux/actions/app.js
@@ -10,7 +10,7 @@ import {
   selectCurrentModal,
 } from 'redux/selectors/app';
 import { doFetchDaemonSettings } from 'redux/actions/settings';
-import { doBalanceSubscribe, doFetchTransactions } from 'redux/actions/wallet';
+import { doBalanceSubscribe } from 'redux/actions/wallet';
 import { doAuthenticate } from 'redux/actions/user';
 import { doFetchFileInfosAndPublishedClaims } from 'redux/actions/file_info';
 import * as MODALS from 'constants/modal_types';
@@ -217,7 +217,6 @@ export function doDaemonReady() {
     dispatch(doBalanceSubscribe());
     dispatch(doFetchFileInfosAndPublishedClaims());
     dispatch(doFetchRewardedContent());
-    dispatch(doFetchTransactions(false));
     if (!selectIsUpgradeSkipped(state)) {
       dispatch(doCheckUpgradeAvailable());
     }