+++ src/renderer/redux/actions/content.js
@@ -5,7 +5,8 @@ import Lbry from 'lbry';
 import Lbryio from 'lbryio';
 import Lbryuri from 'lbryuri';
 import { makeSelectClientSetting } from 'redux/selectors/settings';
-import { selectBalance, selectTransactionItems } from 'redux/selectors/wallet';
+import { selectMyClaimsRaw } from 'redux/selectors/claims';
+import { selectBalance } from 'redux/selectors/wallet';
 import {
   makeSelectFileInfoForUri,
   selectDownloadingByOutpoint,
@@ -482,8 +483,8 @@ export function doPublish(params) {
 export function doAbandonClaim(txid, nout) {
   return function(dispatch, getState) {
     const state = getState();
-    const transactionItems = selectTransactionItems(state);
-    const { claim_id: claimId, claim_name: name } = transactionItems.find(
+    const myClaims = selectMyClaimsRaw(state);
+    const { claim_id: claimId, name } = myClaims.find(
       claim => claim.txid === txid && claim.nout === nout
     );