+++ src/renderer/store.js
@@ -13,6 +13,7 @@ import userReducer from 'redux/reducers/user';
 import walletReducer from 'redux/reducers/wallet';
 import shapeShiftReducer from 'redux/reducers/shape_shift';
 import subscriptionsReducer from 'redux/reducers/subscriptions';
+import videoReducer from 'redux/reducers/video';
 import { persistStore, autoRehydrate } from 'redux-persist';
 import createCompressor from 'redux-persist-transform-compress';
 import createFilter from 'redux-persist-transform-filter';
@@ -63,6 +64,7 @@ const reducers = combineReducers({
   user: userReducer,
   shapeShift: shapeShiftReducer,
   subscriptions: subscriptionsReducer,
+  video: videoReducer,
 });
 
 const bulkThunk = createBulkThunkMiddleware();